Got marsMissions separated and exported to server.js
