+++ models/marsMissions.js
@@ -1,0 +1,39 @@
+const marsMissions = [
+    {
+      name: "Curiosity",
+      launchDate: "26 Nov 2011",
+      operator: "NASA",
+      missionType: "Rover",
+      img: ""
+    },
+    {
+      name: "Opportunity",
+      launchDate: "8 Jul 2003",
+      operator: "NASA",
+      missionType: "Rover",
+      img: ""
+    },
+    {
+      name: "Spirit",
+      launchDate: "10 Jun 2003",
+      operator: "NASA",
+      missionType: "Rover",
+      img: ""
+    },
+    {
+      name: "Sojourner",
+      launchDate: "4 Dec 1996",
+      operator: "NASA",
+      missionType: "Rover",
+      img: ""
+    },
+    {
+      name: "Rosetta",
+      launchDate: "2 Mar 2004",
+      operator: "ESA",
+      missionType: "Gravity Assist",
+      img: ""
+    }
+  ];
+
+  module.exports = marsMissions;
