+++ server.js
@@ -1,6 +1,7 @@
 // DEPENDENCIES
 const express = require('express');
 const app = express();
+const marsMissions = require('./models/marsMissions');
 
 // run `npm install` to install dependencies in package.json
 
@@ -32,43 +33,16 @@ const port = 3000;
 // DATA - move this to a file called marsMissions.js inside of a models folder
 // remember to export (module.exports)
 // remember to require it in the server
-const marsMissions = [
-  {
-    name: "Curiosity",
-    launchDate: "26 Nov 2011",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Opportunity",
-    launchDate: "8 Jul 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Spirit",
-    launchDate: "10 Jun 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Sojourner",
-    launchDate: "4 Dec 1996",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Rosetta",
-    launchDate: "2 Mar 2004",
-    operator: "ESA",
-    missionType: "Gravity Assist",
-    img: ""
-  }
-];
+console.log(marsMissions);
+
+app.get('/', (request, response) =>{
+  console.log("Hi there!");
+  response.send(marsMissions);
+
+})
+  
+
+
 
 
 // LISTENER
